+++ R/summary.R
@@ -78,3 +78,63 @@ print.summary.aftsrr <- function(x, ...){
   }
 }
 
+format.perc <- function (probs, digits) 
+  paste(format(100 * probs, trim = TRUE, scientific = FALSE, digits = digits), "%")
+
+#' @export
+confint.aftsrr <- function(object, parm, level = 0.95, ...) {
+  cf <- coef(object)
+  pnames <- names(cf)
+  object$covmat <- object$covmat[!is.na(object$covmat)]
+  if (!length(object$covmat))
+    stop("Missing covariance-variance estimate.")
+  vnames <- names(object$covmat)
+  ses <- lapply(object$covmat, function(e) {
+    ses <- sqrt(diag(e))
+    names(ses) <- pnames
+    return(ses)
+  })
+  if (is.matrix(cf)) 
+    cf <- setNames(as.vector(cf), pnames)
+  if (missing(parm)) 
+    parm <- pnames
+  else if (is.numeric(parm)) 
+    parm <- pnames[parm]
+  a <- (1 - level)/2
+  a <- c(a, 1 - a) 
+  fac <- qnorm(a)
+  ciList <- lapply(ses, function(e) {
+    pct <- format.perc(a, 3)
+    ci <- array(NA_real_, dim = c(length(parm), 2L), dimnames = list(parm, pct))
+    ci[] <- cf[parm] + e[parm] %o% fac
+    ci
+  })
+  for (i in 1:length(ciList)) {
+    cat("\n")
+    cat("Variance Estimator:", as.character(names(ciList)[i]))
+    cat("\n")
+    print(ciList[[i]])
+  }
+  invisible(ciList)  
+}
+
+#' @export
+confint.aftgee <- function(object, parm, level = 0.95, ...) {
+  cf <- coef(object)
+  pnames <- names(cf)
+  ses <- sqrt(diag(vcov(object)))
+  names(ses) <- pnames
+  if (is.matrix(cf)) 
+    cf <- setNames(as.vector(cf), pnames)
+  if (missing(parm)) 
+    parm <- pnames
+  else if (is.numeric(parm)) 
+    parm <- pnames[parm]
+  a <- (1 - level)/2
+  a <- c(a, 1 - a) 
+  fac <- qnorm(a)
+  pct <- format.perc(a, 3)
+  ci <- array(NA_real_, dim = c(length(parm), 2L), dimnames = list(parm, pct))
+  ci[] <- cf[parm] + ses[parm] %o% fac
+  ci
+}